// apps/web/src/types/docs-types.ts
export enum ClientDocsPanel {
    OVERVIEW = 'OVERVIEW',
    HOW_IT_WORKS = 'HOW_IT_WORKS',
    GETTING_STARTED = 'GETTING_STARTED',
    EDITING_AND_UNDERSTANDING = 'EDITING_AND_UNDERSTANDING',
    BUILDING_AND_TESTING = 'BUILDING_AND_TESTING',
    DEPLOYMENT = 'DEPLOYMENT',
    EXPORTING = 'EXPORTING',
    SANDBOX = 'SANDBOX',
    SANDBOX_INTRO = 'SANDBOX_INTRO',
    SANDBOX_CONFIGURATION = 'SANDBOX_CONFIGURATION',
    SANDBOX_USAGE = 'SANDBOX_USAGE',
}